fix: wrong type for MAX_BUFFER_SIZE

+++ components/balboa/balboa.h
@@ -24,7 +24,7 @@ class BalboaClimate;
 class BalboaSwitch;
 class BalboaLightSelect;
 
-const uint8_t MAX_BUFFER_SIZE = 256;
+const size_t MAX_BUFFER_SIZE = 256;
 
 const uint8_t MSME = 0x7E;
 const uint8_t CHANNEL_MULTICAST = 0xFE;
